requirements.txt
pandas
numpy
pytesseract
pillow